% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{check_has_G}
\alias{check_has_G}
\title{Check if a dlm block has the treatment as covariate}
\usage{
check_has_G(formula, name.G, inside_dlm = FALSE)
}
\arguments{
\item{formula}{A formula describing the model passed to the fit.surr function.}

\item{name.G}{The name of the surrogate}
}
\description{
Check if a dlm block has the treatment as covariate
}
\keyword{internal}
