% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{check_is_dlm_block}
\alias{check_is_dlm_block}
\title{Check if a formula term is a dlm block}
\usage{
check_is_dlm_block(term)
}
\arguments{
\item{term}{A string.}
}
\description{
Check if a formula term is a dlm block
}
\keyword{internal}
