% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{has_call}
\alias{has_call}
\title{Test whether an expression contains a function call}
\usage{
has_call(expr, fun = "s")
}
\arguments{
\item{expr}{An R language object.}

\item{fun}{A character string giving the function name to look for.}
}
\value{
\code{TRUE} if \code{expr} contains a call to \code{fun}, otherwise \code{FALSE}.
}
\description{
Recursively checks whether \code{expr} contains a call to \code{fun}.
}
\keyword{internal}
