% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{lagged}
\alias{lagged}
\title{Compute lagged values of a vector}
\usage{
lagged(x, k = 1)
}
\arguments{
\item{x}{A vector to be lagged.}

\item{k}{A non-negative integer giving the lag order.}
}
\value{
A vector of the same length as \code{x}, with the first \code{k} values set to \code{0} and the remaining values taken from \code{x} shifted by \code{k} positions.
}
\description{
Returns a lagged version of \code{x} by shifting its values forward by \code{k} positions and padding the first \code{k} entries with zeros.
}
\details{
This function is intended for use in model formulas when delayed effects of a predictor should be included explicitly.
}
\examples{
x <- 1:5

lagged(x)
lagged(x, k = 2)

}
