% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fitted_onlinesurr}
\alias{print.fitted_onlinesurr}
\title{print.fitted_onlinesurr}
\usage{
\method{print}{fitted_onlinesurr}(x, ...)
}
\arguments{
\item{x}{A fitted_onlinesurr object.}

\item{...}{Arguments passed to summary.fitted_onlinesurr}
}
\value{
No return value, called to print a summary of the fitted kDGLM model.
}
\description{
This method is wrapper for the summary.fitted_onlinesurr method.
}
\keyword{internal}
