% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{time_homo_test}
\alias{time_homo_test}
\title{Test time-homogeneity of the PTE}
\usage{
time_homo_test(model, signif.level = 0.05, N.boots = 50000)
}
\arguments{
\item{model}{A fitted object of class \code{"fitted_onlinesurr"}, typically returned by \code{fit.surr}. Must contain \code{$T}, \code{$n.fixed}, and the elements \code{$Marginal} and \code{$Conditional} with \code{point} and \code{smp} components.}

\item{signif.level}{Numeric in (0,1) giving the test significance level used to form the critical value from the bootstrap distribution. Default is \code{0.05}.}

\item{N.boots}{Integer number of Monte Carlo draws used to approximate the null distribution of the max standardized deviation statistic and to compute the p-value. Default is \code{50000}.}
}
\value{
A named list with:
  \itemize{
    \item \code{T}: the observed test statistic (maximum absolute standardized deviation).
    \item \code{T.crit}: the 1-signif.level critical value.
    \item \code{p.value}: the Monte Carlo p-value \code{mean(T_null > T_obs)}.
  }
}
\description{
Tests the null hypothesis that the LPTE is constant over time. The test is based on the difference between the conditional and marginal treatment-effect trajectories implied by a fitted \code{"fitted_onlinesurr"} object, standardized by an estimated covariance, and uses a max-type statistic to control the family wise error across time points.
}
\details{
See \insertCite{santos2026causalframeworkevaluatingjointly}{OnlineSurr} for the theoretical details about this test.



Notes:
\itemize{
  \item The function assumes the first \code{T} time-specific treatment-effect parameters are stored contiguously at the beginning of \code{model$Marginal$point} and \code{model$Conditional$point} (and similarly for \code{smp}). It uses the index \code{1:(n.fixed)} as implemented in the code: \code{1:(T + n.fixed - T)}.
  \item \code{N.boots} here is a Monte Carlo size for the null simulation (distinct from the bootstrap size used when fitting \code{model}).
}
}
\examples{
\dontrun{
fit <- fit.surr(y ~ 1,
  id = id, surrogate = ~ s1 + s2, treat = trt,
  data = dat, time = time, N.boots = 2000
)

time_homo_test(fit, signif.level = 0.05, N.boots = 50000)
}
}
